// frontend/src/api/generated/models/CreatePromptCategoryDto.ts
//@ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * c4 GenAI Suite
 * c4 GenAI Suite
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreatePromptCategoryDto
 */
export interface CreatePromptCategoryDto {
    /**
     * The name of the prompt category
     * @type {string}
     * @memberof CreatePromptCategoryDto
     */
    name: string;
    /**
     * Optional description of the category
     * @type {string}
     * @memberof CreatePromptCategoryDto
     */
    description?: string;
    /**
     * Hex color code for the category
     * @type {string}
     * @memberof CreatePromptCategoryDto
     */
    color?: string;
    /**
     * Sort order for displaying categories
     * @type {number}
     * @memberof CreatePromptCategoryDto
     */
    sortOrder?: number;
}

/**
 * Check if a given object implements the CreatePromptCategoryDto interface.
 */
export function instanceOfCreatePromptCategoryDto(value: object): value is CreatePromptCategoryDto {
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function CreatePromptCategoryDtoFromJSON(json: any): CreatePromptCategoryDto {
    return CreatePromptCategoryDtoFromJSONTyped(json, false);
}

export function CreatePromptCategoryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreatePromptCategoryDto {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'color': json['color'] == null ? undefined : json['color'],
        'sortOrder': json['sortOrder'] == null ? undefined : json['sortOrder'],
    };
}

export function CreatePromptCategoryDtoToJSON(value?: CreatePromptCategoryDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'name': value['name'],
        'description': value['description'],
        'color': value['color'],
        'sortOrder': value['sortOrder'],
    };
}